Refactor factory
